+++ src/steps/resource-manager/storage/client.ts
@@ -3,12 +3,58 @@ import {
   BlobContainer,
   FileShare,
   StorageAccount,
+  StorageQueue,
+  Kind,
 } from '@azure/arm-storage/esm/models';
+
+interface ListStorageAccountResourcesEndpoint extends ListResourcesEndpoint {
+  list<ListResponseType>(
+    resourceGroupName: string,
+    accountName: string,
+  ): Promise<ListResponseType>;
+  listNext<ListResponseType>(nextLink: string): Promise<ListResponseType>;
+}
+
+interface IterateAllStorageAccountResourcesOptions<ResourceType>
+  extends IterateAllResourcesOptions<any, ResourceType> {
+  resourceGroupName: string;
+  accountName: string;
+  serviceClient: StorageManagementClient;
+  resourceEndpoint: ListStorageAccountResourcesEndpoint;
+}
+
+async function iterateAllStorageAccountResources<ResourceType>({
+  serviceClient,
+  resourceEndpoint,
+  resourceDescription,
+  callback,
+  logger,
+  resourceGroupName,
+  accountName,
+}: IterateAllStorageAccountResourcesOptions<ResourceType>) {
+  return iterateAllResources({
+    logger,
+    serviceClient,
+    resourceEndpoint: {
+      list: async () => {
+        return resourceEndpoint.list(resourceGroupName, accountName);
+      },
+      listNext: /* istanbul ignore next: testing iteration might be difficult */ async (
+        nextLink: string,
+      ) => {
+        return resourceEndpoint.listNext(nextLink);
+      },
+    } as ListResourcesEndpoint,
+    resourceDescription,
+    callback,
+  });
+}
 
 import {
   Client,
   iterateAllResources,
   ListResourcesEndpoint,
+  IterateAllResourcesOptions,
 } from '../../../azure/resource-manager/client';
 import { resourceGroupName } from '../../../azure/utils';
 
@@ -55,6 +101,30 @@ export class StorageClient extends Client {
       resourceDescription: 'storage.blobContainers',
       callback,
     });
+  }
+
+  /* eslint-disable @typescript-eslint/no-non-null-assertion */
+  public async iterateQueues(
+    storageAccount: { name: string; id: string; kind: Kind },
+    callback: (e: StorageQueue) => void | Promise<void>,
+  ): Promise<void> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      StorageManagementClient,
+    );
+    const resourceGroup = resourceGroupName(storageAccount.id, true)!;
+    const accountName = storageAccount.name!;
+
+    if ((['Storage', 'StorageV2'] as Kind[]).includes(storageAccount.kind)) {
+      return iterateAllStorageAccountResources({
+        logger: this.logger,
+        serviceClient,
+        resourceEndpoint: serviceClient.queue,
+        resourceGroupName: resourceGroup,
+        accountName,
+        resourceDescription: 'storage.queues',
+        callback,
+      });
+    }
   }
 
   public async iterateFileShares(